added file_size parameter in table_cache get function; refactored btree iterator;

+++ db/version_set.h
@@ -144,7 +144,7 @@ class Version {
 
   // List of files per level
   std::vector<FileMetaData*> files_[config::kNumLevels];
-  std::unordered_map<uint64_t, std::pair<FileMetaData*, uint64_t>> metadata_map_;
+  std::unordered_map<uint64_t, std::pair<FileMetaData*, int>> metadata_map_;
 
   // Next file to compact based on seek stats.
   FileMetaData* file_to_compact_;
